DVT-#100 레이아웃 수정
- 프로필 카드 전체에 cursor: pointer를 제거하고 이미지에만 추가함 (좋아요를 위해)
- theme에 코스별 색상 추가
